Administrateur:     - Créer utilisateur (auth, isAdmin) -- DONE     - Modifier utilisateur (auth, isAdmin) -- DONE     - Gérer les salles (auth, isAdmin) -- DONE     - Gérer les structures (auth, isAdmin) -- DONE     - Générer le fichier xml des utilisateurs (auth, isAdmin)     - Consulter salle (auth) -- DONE
Responsable:
    - Consulter les réservations d'aujourd'hui -- DONE
    - Ajouter réservation (auth, canReserve) -- DONE
    - Modifier réservation (auth, canReserve) -- TO DO
    - Supprimer Réservation (auth, canReserve) -- DONE
    - Consulter Salle (auth) -- DONE

Secretariat:

    - Consulter les réservations (auth) -- DONE
    - Ajouter réservation (auth, canReserve) -- DONE
    - Suprimer réservation (auth, canReserve) -- DONE
    - Modifier réservation (auth, canReserve) -- DONE
    - Valider réservation (auth, canReserve) -- TO DO
    - Consulter Salle (auth) -- DONE

Utilisateur:
    - Consulter réservations des salles (auth) -- DONE
    - Modifier Profile -- DONE

Non Authentifié:
    - Consulter réservations du jour -- DONE

+++ database/migrations/2022_11_24_092034_create_reservation_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('utilisateur_id');
             $table->unsignedBigInteger('salle_id');
             $table->dateTime('date_reservation');
-            $table->integer('reservation_periode');
+            $table->unsignedBigInteger('reservation_periode');
             $table->string('reservation_commentaire')->nullable();
             $table->unsignedBigInteger('reservation_statut'); // provisoire, confirmé, annulé
